Bug 19145 This issue was related to the event timestamp buffer overflowing without the counter being reset. increased the size of the rollover counter and changed the messages types for debugging
(cherry picked from commit 40d45c1268a4fc87dbf93df3b8e319f94288d5d5)

+++ common/source/EventTask.c
@@ -48,7 +48,7 @@ typedef struct EventTaskTag {
 	uint32_t timeStampBuffer[TOTAL_NUMBER_EVENTS];
 	uint32_t advertisingEvent;
 	uint32_t eventID;
-	uint8_t eventBufferRollover;
+	uint32_t eventBufferRollover;
 } EventTaskObj_t;
 /******************************************************************************/
 /* Local Data Definitions                                                     */
